platform-demos: Edited the Hello World java script tutorial for clarity
The original tutorial had improper grammar and spelling, and left out
one line of code in the step-by-step examples. It also told the reader
about making the resulting script executable long before the script was
finished and saved to disc. These edits are intended to make the tutorial
qeasier to read and complete, and less likely to stymie the reader.