New directory for beginners tutorials
New directory for beginners tutorials, to be written starting from the
platform-demos examples and GTK+ tutorials.

I have also renamed the pages beginner.<language>.page as gtk-beginner.
<language>.page - except for python, where the starting point of the tutorial
was tutorial.py.page. The latter page has been rewritten, dividing the tutorial
into sections that can be browsed more easily.

I have started rewriting the GTK+ tutorial for Python. window.py.page is
complete; the pages for GtkApplicationWindow and image are in progress. The
rest is to be done.